Reorganize tests for Netv6Addr and split impls into multiple files
