+++ src/netv6addr/from.rs
@@ -1,0 +1,1 @@
+mod ipv6addr;
